auto submit

+++ exhibit/handlers/order.py
@@ -158,7 +158,7 @@ def list(request):
     orderType = "0"
   # 处理duration
   for order in orders:
-    order.price = order.price / 10.0
+    order.price = int(order.price / 10.0)
     if order.order_type == "1":
       order.duration = order.end_time - order.start_time
       order.duration = order.duration.seconds / 3600.0
